Moves game event type package into the internal directory to avoid confusion with SDK events
