+++ simple-game-server-go/internal/event/event.go
@@ -3,12 +3,12 @@ package event
 import "github.com/Unity-Technologies/multiplay-examples/simple-game-server-go/pkg/config"
 
 type (
-	// Type is a type of Multiplay game server lifecycle event.
-	Type int
+	// LifecycleEventType is a type of Multiplay game server lifecycle event.
+	LifecycleEventType int
 
-	// Event represents a Multiplay game server lifecycle event.
-	Event struct {
-		Type           Type
+	// LifecycleEvent represents a Multiplay game server lifecycle event.
+	LifecycleEvent struct {
+		EventType      LifecycleEventType
 		AllocationUUID string
 		Config         *config.Config
 	}
@@ -17,7 +17,7 @@ type (
 const (
 	// Allocated indicates that a matchmaker has requested a game server from
 	// Multiplay and this one has been chosen to host a match.
-	Allocated = Type(iota)
+	Allocated = LifecycleEventType(iota)
 
 	// Deallocated indicates that the matchmaker no longer requires this game
 	// server to host a match.
