+++ simple-game-server-go/internal/game/game.go
@@ -8,8 +8,8 @@ import (
 	"sync"
 	"time"
 
+	"github.com/Unity-Technologies/multiplay-examples/simple-game-server-go/internal/event"
 	"github.com/Unity-Technologies/multiplay-examples/simple-game-server-go/pkg/config"
-	"github.com/Unity-Technologies/multiplay-examples/simple-game-server-go/pkg/event"
 	"github.com/Unity-Technologies/multiplay-examples/simple-game-server-go/pkg/proto"
 	"github.com/Unity-Technologies/multiplay-examples/simple-game-server-go/pkg/sdkclient"
 	"github.com/hashicorp/go-multierror"
@@ -28,7 +28,7 @@ type (
 		clients sync.Map
 
 		// gameEvents is a channel of game events, for example allocated / deallocated
-		gameEvents chan event.Event
+		gameEvents chan event.LifecycleEvent
 
 		// gameBind is a TCP listener representing a fake game server
 		gameBind *net.TCPListener
@@ -45,10 +45,6 @@ type (
 
 		// queryBind is a UDP endpoint which responds to game queries
 		queryBind *udpBinding
-
-		// internalEventProcessorReady is a channel that, when written to,
-		// indicates that the internal event processor is ready.
-		internalEventProcessorReady chan struct{}
 
 		// queryPort is the port number the game query server will listen on
 		queryPort uint
@@ -80,14 +76,13 @@ type (
 // New creates a new game, configured with the provided configuration file.
 func New(logger *logrus.Entry, configPath string, port, queryPort uint, httpClient *http.Client) (*Game, error) {
 	g := &Game{
-		cfgFile:                     configPath,
-		gameEvents:                  make(chan event.Event, 1),
-		logger:                      logger,
-		internalEventProcessorReady: make(chan struct{}, 1),
-		done:                        make(chan struct{}, 1),
-		port:                        port,
-		queryPort:                   queryPort,
-		httpClient:                  httpClient,
+		cfgFile:    configPath,
+		gameEvents: make(chan event.LifecycleEvent, 1),
+		logger:     logger,
+		done:       make(chan struct{}, 1),
+		port:       port,
+		queryPort:  queryPort,
+		httpClient: httpClient,
 	}
 
 	return g, nil
@@ -143,7 +138,7 @@ func (g *Game) Stop() error {
 		g.queryBind.Close()
 	}
 
-	g.gameEvents <- event.Event{Type: event.Deallocated}
+	g.gameEvents <- event.LifecycleEvent{EventType: event.Deallocated}
 	close(g.done)
 	close(g.gameEvents)
 	g.wg.Wait()
